feat(ring): Implement SPSC ring buffer and wait budget - Added `Spsc` struct for single-producer/single-consumer ring buffer with power-of-two capacity and optimized for delegation patterns. - Introduced `WaitBudget` for lightweight wait management, allowing bounded spin and yield without parking. - Updated dependencies in `Cargo.toml` and removed unused packages from `Cargo.lock`. - Refactored `trustee` module to utilize the new ring buffer implementation.
