+++ Cargo.toml
@@ -12,23 +12,10 @@ debug = true
 [dependencies]
 may = { version = "0.3.51" }
 smallvec = "1.15.1"
-crossbeam-queue = "0.3.12"
 libc = "0.2"
-branches = "0.3.0"
 
 [features]
 default = []
-serde = ["dep:serde", "dep:bincode"]
-
-[dependencies.bincode]
-version = "2.0.1"
-optional = true
-features = ["serde"]
-
-[dependencies.serde]
-version = "1.0"
-features = ["derive"]
-optional = true
 
 [dev-dependencies]
 quickcheck = "1.0.3"
