+++ src/fiber.rs
@@ -3,12 +3,15 @@
 pub struct DelegatedScopeGuard {
     _priv: (),
 }
+
+/// Enter a delegated scope. Used for testing.
 impl DelegatedScopeGuard {
     #[inline]
     pub fn enter() -> Self {
         Self { _priv: () }
     }
 }
+
 impl Drop for DelegatedScopeGuard {
     #[inline]
     fn drop(&mut self) {}
