+++ src/ring/spsc.rs
@@ -6,7 +6,7 @@ use core::cell::UnsafeCell;
 use core::mem::MaybeUninit;
 use core::sync::atomic::{AtomicUsize, Ordering};
 
-use branches::{likely, unlikely};
+use crate::ring::wait::WaitBudget;
 
 const CACHELINE: usize = 64;
 
@@ -88,7 +88,7 @@ impl<T> Spsc<T> {
     pub fn try_pop(&self) -> Option<T> {
         let tail = self.tail.val.load(Ordering::Acquire);
         let head = self.head.val.load(Ordering::Relaxed);
-        if unlikely(head == tail) {
+        if head == tail {
             return None;
         }
         let idx = head & self.mask;
@@ -137,47 +137,6 @@ impl<T> Drop for Spsc<T> {
                 std::ptr::drop_in_place((*self.buf[idx].get()).as_mut_ptr());
             }
             head = head.wrapping_add(1);
-        }
-    }
-}
-
-/// Lightweight wait budget: bounded spin then yield; never parks by default.
-#[derive(Copy, Clone)]
-pub struct WaitBudget {
-    spins: u32,
-    yields: u32,
-    spin_cap: u32,
-    yield_cap: u32,
-}
-
-impl WaitBudget {
-    #[inline]
-    pub fn hot() -> Self {
-        Self {
-            spins: 0,
-            yields: 0,
-            spin_cap: 128,
-            yield_cap: 8,
-        }
-    }
-
-    #[inline]
-    pub fn reset(&mut self) {
-        self.spins = 0;
-        self.yields = 0;
-    }
-
-    #[inline]
-    pub fn step(&mut self) {
-        if likely(self.spins < self.spin_cap) {
-            core::hint::spin_loop();
-            self.spins += 1;
-        } else if likely(self.yields < self.yield_cap) {
-            std::thread::yield_now();
-            self.yields += 1;
-        } else {
-            // Stay hot without parking to avoid scheduler-induced latency.
-            core::hint::spin_loop();
         }
     }
 }
